% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpFlowData-package.R
\docType{package}
\name{sgpFlowData-package}
\alias{sgpFlowData}
\alias{sgpFlowData-package}
\title{sgpFlowData: Exemplar Data Sets for sgpFlow Analyses}
\description{
The `sgpFlowData` package provides exemplar data sets for sgpFlow testing and analyses.
}
\details{
\tabular{ll}{
  Package: \tab sgpFlowData \cr
  Type: \tab Package \cr
  Version: \tab 0.0-0.1 \cr
  Date: \tab 2025-5-23 \cr
  License: \tab MIT \cr
  LazyLoad: \tab yes \cr
}
 
**Data Sets:**
- **sgpFlowData_LONG:** Long formatted data set for sgpFlow analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CenterForAssessment.github.io/sgpFlowData/}
  \item \url{https://github.com/CenterForAssessment/sgpFlowData/,}
  \item Report bugs at \url{https://github.com/CenterForAssessment/sgpFlowData/issues/}
}

}
\author{
\strong{Maintainer}: Damian W. Betebenner \email{dbetebenner@nciea.org}

Authors:
\itemize{
  \item Adam R. Van Iwaarden \email{avaniwaarden@nciea.com}
}

}
\keyword{package}
