% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpFlowData-package.R
\docType{data}
\name{sgpFlowData_LONG}
\alias{sgpFlowData_LONG}
\title{State-Specific Meta-Data for SGP Flow Analyses}
\format{
A LONG formatted longitudinal data set consisting of annual student summative assessment data.
}
\usage{
sgpFlowData_LONG
}
\description{
The `sgpFlowDataStateData` object contains state-specific configurations and meta-data required for SGP analysis. It facilitates tailored analyses by providing:

- Conditional Standard Error of Measurement (CSEM) values for score perturbation
- Achievement level data for state-specific assessments
- Projection spline matrices for modeling student growth

**Contents:**
- `Achievement`: A list containing CSEM values and metadata for state-specific achievement.
- `ProjectionMatrices`: State-specific projection spline matrices used for growth modeling.

The `sgpFlowDataStateData` object is preloaded into the package and is automatically accessed by relevant functions.
}
\examples{
\dontrun{
# View the first few rows of the data set
head(sgpFlowData_LONG)

# View the structure of the data set
str(sgpFlowData_LONG)
}
}
\keyword{datasets}
